Config: make all fields required
This change removes Semigroup and Monoid instances and updates Config to
make all fields required. In practice, they're all required anyway, and
there were previously lots of places where the lack of some settings
caused just-in-time attempts to set them by loading a default
configuration. This patch cleans all that up by pushing that concern to
the very top level, at mkVty, after which point the required values are
always available to all the code to be called later.