Make vty-unix-build-width-table more robust
Previously, the tool could fail if the terminal failed to recognize the getCursorPosition escape sequence. This failure mode is uncommon in general but seems to happen occasional in my testing.

The new mitigating logic is to retry a few times before defaulting to width 1 and moving on. A warning is printed to the terminal when this happens.